+++ packages/api/endpoints/granules.js
@@ -2,8 +2,8 @@
 
 const router = require('express-promise-router')();
 const isBoolean = require('lodash/isBoolean');
+const { v4: uuidv4 } = require('uuid');
 
-const asyncOperations = require('@cumulus/async-operations');
 const Logger = require('@cumulus/logger');
 const { deconstructCollectionId } = require('@cumulus/message/Collections');
 const {
@@ -26,6 +26,7 @@ const {
   deleteGranuleAndFiles,
 } = require('../src/lib/granule-delete');
 const { chooseTargetExecution } = require('../lib/executions');
+const startAsyncOperation = require('../lib/startAsyncOperation');
 const {
   createGranuleFromApi,
   updateGranuleFromApi,
@@ -34,7 +35,6 @@ const {
 } = require('../lib/writeRecords/write-granules');
 const { asyncOperationEndpointErrorHandler } = require('../app/middleware');
 const { errorify } = require('../lib/utils');
-const AsyncOperation = require('../models/async-operation');
 const { moveGranule, getFilesExistingAtLocation } = require('../lib/granules');
 const { reingestGranule, applyWorkflow } = require('../lib/ingest');
 const { unpublishGranule } = require('../lib/granule-remove-from-cmr');
@@ -520,10 +520,6 @@ async function bulkOperations(req, res) {
   if (!payload.workflowName) {
     return res.boom.badRequest('workflowName is required.');
   }
-  const stackName = process.env.stackName;
-  const systemBucket = process.env.system_bucket;
-  // TODO -- Phase 3 -- remove this env variable setting when we update the async model
-  const tableName = process.env.AsyncOperationsTable;
 
   let description;
   if (payload.query) {
@@ -534,9 +530,9 @@ async function bulkOperations(req, res) {
     description = `Bulk run on ${payload.workflowName}`;
   }
 
-  const asyncOperation = await asyncOperations.startAsyncOperation({
-    asyncOperationTaskDefinition: process.env.AsyncOperationTaskDefinition,
-    cluster: process.env.EcsCluster,
+  const asyncOperationId = uuidv4();
+  const asyncOperationEvent = {
+    asyncOperationId,
     callerLambdaName: getFunctionNameFromRequestContext(req),
     lambdaName: process.env.BulkOperationLambda,
     description,
@@ -556,14 +552,11 @@ async function bulkOperations(req, res) {
         METRICS_ES_PASS: process.env.METRICS_ES_PASS,
       },
     },
-    esHost: process.env.ES_HOST,
-    stackName,
-    systemBucket,
-    dynamoTableName: tableName,
-    knexConfig: process.env,
-  }, AsyncOperation);
+  };
 
-  return res.status(202).send(asyncOperation);
+  log.debug(`About to invoke lambda to start async operation ${asyncOperationId}`);
+  await startAsyncOperation.invokeStartAsyncOperationLambda(asyncOperationEvent);
+  return res.status(202).send({ id: asyncOperationId });
 }
 
 /**
@@ -580,13 +573,9 @@ async function bulkDelete(req, res) {
     return res.boom.badRequest('forceRemoveFromCmr must be a boolean value');
   }
 
-  const stackName = process.env.stackName;
-  const systemBucket = process.env.system_bucket;
-  const tableName = process.env.AsyncOperationsTable;
-
-  const asyncOperation = await asyncOperations.startAsyncOperation({
-    asyncOperationTaskDefinition: process.env.AsyncOperationTaskDefinition,
-    cluster: process.env.EcsCluster,
+  const asyncOperationId = uuidv4();
+  const asyncOperationEvent = {
+    asyncOperationId,
     callerLambdaName: getFunctionNameFromRequestContext(req),
     lambdaName: process.env.BulkOperationLambda,
     description: 'Bulk granule deletion',
@@ -614,28 +603,22 @@ async function bulkDelete(req, res) {
         ES_HOST: process.env.ES_HOST,
       },
     },
-    stackName,
-    systemBucket,
-    dynamoTableName: tableName,
-    knexConfig: process.env,
-  }, AsyncOperation);
+  };
 
-  return res.status(202).send(asyncOperation);
+  log.debug(`About to invoke lambda to start async operation ${asyncOperationId}`);
+  await startAsyncOperation.invokeStartAsyncOperationLambda(asyncOperationEvent);
+  return res.status(202).send({ id: asyncOperationId });
 }
 
 async function bulkReingest(req, res) {
   const payload = req.body;
-  const stackName = process.env.stackName;
-  const systemBucket = process.env.system_bucket;
-  const tableName = process.env.AsyncOperationsTable;
-
   const numOfGranules = (payload.query && payload.query.size)
     || (payload.ids && payload.ids.length);
   const description = `Bulk granule reingest run on ${numOfGranules || ''} granules`;
 
-  const asyncOperation = await asyncOperations.startAsyncOperation({
-    asyncOperationTaskDefinition: process.env.AsyncOperationTaskDefinition,
-    cluster: process.env.EcsCluster,
+  const asyncOperationId = uuidv4();
+  const asyncOperationEvent = {
+    asyncOperationId,
     callerLambdaName: getFunctionNameFromRequestContext(req),
     lambdaName: process.env.BulkOperationLambda,
     description,
@@ -655,14 +638,11 @@ async function bulkReingest(req, res) {
         METRICS_ES_PASS: process.env.METRICS_ES_PASS,
       },
     },
-    esHost: process.env.ES_HOST,
-    stackName,
-    systemBucket,
-    dynamoTableName: tableName,
-    knexConfig: process.env,
-  }, AsyncOperation);
+  };
 
-  return res.status(202).send(asyncOperation);
+  log.debug(`About to invoke lambda to start async operation ${asyncOperationId}`);
+  await startAsyncOperation.invokeStartAsyncOperationLambda(asyncOperationEvent);
+  return res.status(202).send({ id: asyncOperationId });
 }
 
 router.get('/:granuleName', get);