+++ packages/api/tests/lambdas/test-orca-backup-reconciliation-report.js
@@ -192,7 +192,11 @@ test.beforeEach(async (t) => {
   esAlias = randomId('esalias');
   esIndex = randomId('esindex');
   process.env.ES_INDEX = esAlias;
-  await bootstrapElasticSearch('fakehost', esIndex, esAlias);
+  await bootstrapElasticSearch({
+    host: 'fakehost',
+    index: esIndex,
+    alias: esAlias,
+  });
   esClient = await Search.es();
 });